コレクションをその場で作成してオカズを追加するフローに対応
