+++ resources/assets/js/user/collections.tsx
@@ -1,161 +1,23 @@
 import React, { useContext, useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter, Link, Outlet, Route, Routes, useNavigate, useParams } from 'react-router-dom';
-import { Button, Form, Modal, ModalProps } from 'react-bootstrap';
-import classNames from 'classnames';
+import { Button } from 'react-bootstrap';
 import { MyProfileContext, useMyProfile } from '../context';
 import { useFetchMyProfile, useFetchCollections, useFetchMyCollections } from '../api';
 import { showToast } from '../tissue';
 import { fetchPostJson, ResponseError } from '../fetch';
-import { FieldError } from '../components/FieldError';
-import { ProgressButton } from '../components/ProgressButton';
 import { Collection } from './collection';
+import {
+    CollectionEditModal,
+    CollectionFormErrors,
+    CollectionFormValidationError,
+    CollectionFormValues,
+} from '../components/collections/CollectionEditModal';
 
 export const CollectionsContext = React.createContext<ReturnType<typeof useFetchCollections> | undefined>(undefined);
 export const MyCollectionsContext = React.createContext<ReturnType<typeof useFetchMyCollections> | undefined>(
     undefined
 );
-
-export type CollectionFormValues = {
-    title: string;
-    is_private: boolean;
-};
-
-export type CollectionFormErrors = {
-    [Property in keyof CollectionFormValues]+?: string[];
-};
-
-export class CollectionFormValidationError extends Error {
-    errors: CollectionFormErrors;
-
-    constructor(errors: CollectionFormErrors, ...rest: any) {
-        super(...rest);
-        if (Error.captureStackTrace) {
-            Error.captureStackTrace(this, CollectionFormValidationError);
-        }
-
-        this.name = 'CollectionFormValidationError';
-        this.errors = errors;
-        Object.setPrototypeOf(this, new.target.prototype);
-    }
-}
-
-interface CollectionEditModalProps extends ModalProps {
-    mode: 'create' | 'edit';
-    initialValues: CollectionFormValues;
-    onSubmit: (values: CollectionFormValues) => Promise<void>;
-}
-
-export const CollectionEditModal: React.FC<CollectionEditModalProps> = ({
-    mode,
-    initialValues,
-    onSubmit,
-    show,
-    onHide,
-    ...rest
-}) => {
-    const [values, setValues] = useState(initialValues);
-    const [errors, setErrors] = useState<CollectionFormErrors>({});
-    const [submitting, setSubmitting] = useState(false);
-
-    useEffect(() => {
-        if (show) {
-            setValues(initialValues);
-            setErrors({});
-        }
-    }, [show]);
-
-    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        setSubmitting(true);
-        try {
-            await onSubmit(values);
-        } catch (e) {
-            if (e instanceof CollectionFormValidationError) {
-                setErrors(e.errors);
-                return;
-            }
-            throw e;
-        } finally {
-            setSubmitting(false);
-        }
-    };
-
-    const handleHide = () => {
-        if (!submitting && onHide) {
-            setValues(initialValues);
-            setErrors({});
-            onHide();
-        }
-    };
-
-    return (
-        <Modal show={show} onHide={handleHide} {...rest}>
-            <form onSubmit={handleSubmit}>
-                <Modal.Header closeButton>
-                    <Modal.Title as="h5">コレクションの{mode === 'create' ? '作成' : '設定'}</Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    <div className="form-row">
-                        <div className="form-group col-sm-12">
-                            <label htmlFor="title">
-                                <span className="oi oi-folder" /> タイトル
-                            </label>
-                            <input
-                                type="text"
-                                id="title"
-                                name="title"
-                                className={classNames({ 'form-control': true, 'is-invalid': errors?.title })}
-                                required
-                                value={values.title}
-                                onChange={(e) => setValues((values) => ({ ...values, title: e.target.value }))}
-                            />
-                            <FieldError name="title" label="タイトル" errors={errors?.title} />
-                        </div>
-                    </div>
-                    <div className="form-row">
-                        <div className="form-group col-sm-12">
-                            <p className="mb-1">
-                                <span className="oi oi-eye" /> 公開設定
-                            </p>
-                            <Form.Check
-                                custom
-                                inline
-                                type="radio"
-                                id="collectionItemVisibilityPublic"
-                                label="公開"
-                                checked={!values.is_private}
-                                onChange={() => setValues((values) => ({ ...values, is_private: false }))}
-                            />
-                            <Form.Check
-                                custom
-                                inline
-                                type="radio"
-                                id="collectionItemVisibilityPrivate"
-                                label="非公開"
-                                className="mt-2"
-                                checked={values.is_private}
-                                onChange={() => setValues((values) => ({ ...values, is_private: true }))}
-                            />
-                        </div>
-                    </div>
-                </Modal.Body>
-                <Modal.Footer>
-                    <Button variant="secondary" disabled={submitting} onClick={handleHide}>
-                        キャンセル
-                    </Button>
-                    <ProgressButton
-                        label={mode === 'create' ? '作成' : '更新'}
-                        inProgress={submitting}
-                        type="submit"
-                        variant="primary"
-                        disabled={submitting}
-                    />
-                </Modal.Footer>
-            </form>
-        </Modal>
-    );
-};
 
 type SidebarItemProps = {
     collection: Tissue.Collection;
@@ -278,7 +140,13 @@ const Collections: React.FC = () => {
             <div className="container">
                 <div className="row">
                     <div className="col-lg-4">
-                        <Sidebar collections={fetchCollections.data} reloadCollections={fetchCollections.reload} />
+                        <Sidebar
+                            collections={fetchCollections.data}
+                            reloadCollections={() => {
+                                fetchMyCollections.reload();
+                                fetchCollections.reload();
+                            }}
+                        />
                     </div>
                     <div className="col-lg-8">
                         {fetchCollections.error?.response?.status === 403 ? (
